type hint for accept() observable method

+++ LDG_TV/SC-Java-LDG-TV/src/main/java/domain/model/variable/MoneyVariable.java
@@ -41,6 +41,6 @@ public class MoneyVariable extends Variable {
 
     @Override
     public void accept(Object o) {
-        this.value.setValue(o);
+        this.value.setValue((Integer) o);
     }
 }
